% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_csv.R
\name{annotate_cas_csv}
\alias{annotate_cas_csv}
\title{Annotates data by InChiKey and SMILES based on CAS numbers}
\usage{
annotate_cas_csv(input_csv, output_csv, cas_column)
}
\arguments{
\item{input_csv}{source CSV file containing CAS numbers}

\item{output_csv}{enriched output CSV file}

\item{cas_column}{identification of column with CAS numbers}
}
\description{
Annotates data by InChiKey and SMILES based on CAS numbers
}
