% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_convert.R
\name{cas_to_inchikey}
\alias{cas_to_inchikey}
\title{CAS to InChiKey converter}
\usage{
cas_to_inchikey(cas_number)
}
\arguments{
\item{cas_number}{given CAS number}
}
\value{
appropriate InChiKey value
}
\description{
CAS to InChiKey converter
}
