% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_convert.R
\name{cas_to_smiles}
\alias{cas_to_smiles}
\title{CAS to SMILES converter}
\usage{
cas_to_smiles(cas_number)
}
\arguments{
\item{cas_number}{given CAS number}
}
\value{
appropriate SMILES value
}
\description{
CAS to SMILES converter
}
