% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_csv.R
\name{annotate_cas_csv}
\alias{annotate_cas_csv}
\title{Annotates data by InChiKey and SMILES based on CAS numbers}
\usage{
annotate_cas_csv(csv_filename, cas_column)
}
\arguments{
\item{csv_filename}{source CSV file containing CAS numbers}

\item{cas_column}{identification of column with CAS numbers}
}
\value{
table extended by InChiKey and SMILES values
}
\description{
Annotates data by InChiKey and SMILES based on CAS numbers
}
